# generated/openapi.json
{
  "components": {
    "schemas": {
      "AggregateCatalogResponse": {
        "additionalProperties": false,
        "description": "One read-only view over source-qualified products without deduplication.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogProduct"
            },
            "title": "Items",
            "type": "array"
          },
          "mode": {
            "const": "multi-readonly",
            "default": "multi-readonly",
            "title": "Mode",
            "type": "string"
          },
          "omitted_count": {
            "minimum": 0.0,
            "title": "Omitted Count",
            "type": "integer"
          },
          "partial": {
            "title": "Partial",
            "type": "boolean"
          },
          "read_only": {
            "const": true,
            "default": true,
            "title": "Read Only",
            "type": "boolean"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/AggregateSourceReport"
            },
            "minItems": 2,
            "title": "Sources",
            "type": "array"
          },
          "state": {
            "$ref": "#/components/schemas/AggregateCatalogState"
          },
          "supplier": {
            "const": "aggregate",
            "default": "aggregate",
            "title": "Supplier",
            "type": "string"
          }
        },
        "required": [
          "state",
          "items",
          "sources",
          "partial",
          "omitted_count"
        ],
        "title": "AggregateCatalogResponse",
        "type": "object"
      },
      "AggregateCatalogState": {
        "description": "Whether the combined view is complete, degraded, empty, or unavailable.",
        "enum": [
          "complete",
          "partial",
          "empty",
          "error"
        ],
        "title": "AggregateCatalogState",
        "type": "string"
      },
      "AggregateSourceReport": {
        "additionalProperties": false,
        "description": "Sanitized status for one source contributing to the combined view.",
        "properties": {
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CatalogError"
              },
              {
                "type": "null"
              }
            ]
          },
          "freshness": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CatalogFreshness"
              },
              {
                "type": "null"
              }
            ]
          },
          "item_count": {
            "minimum": 0.0,
            "title": "Item Count",
            "type": "integer"
          },
          "mode": {
            "enum": [
              "mock",
              "khommo-readonly",
              "vietshare-readonly"
            ],
            "title": "Mode",
            "type": "string"
          },
          "omitted_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Omitted Count",
            "type": "integer"
          },
          "partial": {
            "default": false,
            "title": "Partial",
            "type": "boolean"
          },
          "state": {
            "$ref": "#/components/schemas/CatalogState"
          },
          "supplier": {
            "enum": [
              "khommo",
              "vietshare"
            ],
            "title": "Supplier",
            "type": "string"
          }
        },
        "required": [
          "supplier",
          "mode",
          "state",
          "freshness",
          "error",
          "item_count"
        ],
        "title": "AggregateSourceReport",
        "type": "object"
      },
      "ApprovedMappingApproval": {
        "additionalProperties": false,
        "description": "Evidence that an administrator explicitly approved a mapping.",
        "properties": {
          "approved_at": {
            "format": "date-time",
            "title": "Approved At",
            "type": "string"
          },
          "approved_by_admin_id": {
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Approved By Admin Id",
            "type": "string"
          },
          "status": {
            "const": "approved",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "approved_by_admin_id",
          "approved_at"
        ],
        "title": "ApprovedMappingApproval",
        "type": "object"
      },
      "Capability": {
        "additionalProperties": false,
        "description": "Capability state plus an explanation whenever it is unavailable.",
        "properties": {
          "reason": {
            "anyOf": [
              {
                "minLength": 1,
                "pattern": ".*\\S.*",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "$ref": "#/components/schemas/CapabilityStatus"
          }
        },
        "required": [
          "status",
          "reason"
        ],
        "title": "Capability",
        "type": "object"
      },
      "CapabilityStatus": {
        "description": "Availability of one narrow supplier-facing operation.",
        "enum": [
          "enabled",
          "disabled",
          "unsupported"
        ],
        "title": "CapabilityStatus",
        "type": "string"
      },
      "CatalogDetailFound": {
        "additionalProperties": false,
        "description": "A supported catalog detail read that found a product.",
        "properties": {
          "item": {
            "$ref": "#/components/schemas/CatalogProduct"
          },
          "state": {
            "const": "found",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "item"
        ],
        "title": "CatalogDetailFound",
        "type": "object"
      },
      "CatalogDetailNotFound": {
        "additionalProperties": false,
        "description": "A supported detail read for an unknown normalized product id.",
        "properties": {
          "product_id": {
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Product Id",
            "type": "string"
          },
          "state": {
            "const": "not_found",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "product_id"
        ],
        "title": "CatalogDetailNotFound",
        "type": "object"
      },
      "CatalogDetailResponse": {
        "discriminator": {
          "mapping": {
            "found": "#/components/schemas/CatalogDetailFound",
            "not_found": "#/components/schemas/CatalogDetailNotFound",
            "unsupported": "#/components/schemas/CatalogDetailUnsupported"
          },
          "propertyName": "state"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/CatalogDetailFound"
          },
          {
            "$ref": "#/components/schemas/CatalogDetailNotFound"
          },
          {
            "$ref": "#/components/schemas/CatalogDetailUnsupported"
          }
        ]
      },
      "CatalogDetailUnsupported": {
        "additionalProperties": false,
        "description": "An honest result for a supplier without a verified detail schema.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/CatalogError"
          },
          "product_id": {
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Product Id",
            "type": "string"
          },
          "state": {
            "const": "unsupported",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "product_id",
          "error"
        ],
        "title": "CatalogDetailUnsupported",
        "type": "object"
      },
      "CatalogError": {
        "additionalProperties": false,
        "description": "A client-safe error with an explicit retry decision.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/CatalogErrorCode"
          },
          "message": {
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "CatalogError",
        "type": "object"
      },
      "CatalogErrorCode": {
        "description": "Safe, non-secret catalog failure categories exposed to clients.",
        "enum": [
          "source_unavailable",
          "unsupported"
        ],
        "title": "CatalogErrorCode",
        "type": "string"
      },
      "CatalogFreshness": {
        "additionalProperties": false,
        "description": "Deterministic evidence used to derive fresh versus stale state.",
        "properties": {
          "evaluated_at": {
            "format": "date-time",
            "title": "Evaluated At",
            "type": "string"
          },
          "max_age_seconds": {
            "exclusiveMinimum": 0.0,
            "title": "Max Age Seconds",
            "type": "integer"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/FreshnessStatus"
          }
        },
        "required": [
          "status",
          "observed_at",
          "evaluated_at",
          "max_age_seconds"
        ],
        "title": "CatalogFreshness",
        "type": "object"
      },
      "CatalogListResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/CatalogResponse"
          },
          {
            "$ref": "#/components/schemas/AggregateCatalogResponse"
          }
        ]
      },
      "CatalogProduct": {
        "additionalProperties": false,
        "description": "A normalized product containing one or more distinct variants.",
        "properties": {
          "available_quantity": {
            "minimum": 0.0,
            "title": "Available Quantity",
            "type": "integer"
          },
          "description": {
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Description",
            "type": "string"
          },
          "id": {
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "default": "mock",
            "enum": [
              "mock",
              "khommo-readonly",
              "vietshare-readonly"
            ],
            "title": "Mode",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Name",
            "type": "string"
          },
          "price": {
            "$ref": "#/components/schemas/Money"
          },
          "read_only": {
            "const": true,
            "default": true,
            "title": "Read Only",
            "type": "boolean"
          },
          "supplier": {
            "default": "mock",
            "enum": [
              "mock",
              "khommo",
              "vietshare"
            ],
            "title": "Supplier",
            "type": "string"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/CatalogVariant"
            },
            "minItems": 1,
            "title": "Variants",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "price",
          "available_quantity",
          "variants"
        ],
        "title": "CatalogProduct",
        "type": "object"
      },
      "CatalogResponse": {
        "additionalProperties": false,
        "description": "Catalog envelope with internally consistent freshness and error state.",
        "properties": {
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CatalogError"
              },
              {
                "type": "null"
              }
            ]
          },
          "freshness": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CatalogFreshness"
              },
              {
                "type": "null"
              }
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogProduct"
            },
            "title": "Items",
            "type": "array"
          },
          "mode": {
            "enum": [
              "mock",
              "khommo-readonly",
              "vietshare-readonly"
            ],
            "title": "Mode",
            "type": "string"
          },
          "omitted_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Omitted Count",
            "type": "integer"
          },
          "partial": {
            "default": false,
            "title": "Partial",
            "type": "boolean"
          },
          "read_only": {
            "const": true,
            "default": true,
            "title": "Read Only",
            "type": "boolean"
          },
          "state": {
            "$ref": "#/components/schemas/CatalogState"
          },
          "supplier": {
            "default": "mock",
            "enum": [
              "mock",
              "khommo",
              "vietshare"
            ],
            "title": "Supplier",
            "type": "string"
          }
        },
        "required": [
          "mode",
          "state",
          "freshness",
          "items",
          "error"
        ],
        "title": "CatalogResponse",
        "type": "object"
      },
...
      "CatalogState": {
        "description": "Complete UI states represented by the catalog envelope.",
        "enum": [
          "fresh",
          "stale",
          "empty",
          "error"
        ],
        "title": "CatalogState",
        "type": "string"
      },
      "CatalogVariant": {
        "additionalProperties": false,
        "description": "One normalized variant and its exact supplier mapping.",
        "properties": {
          "available_quantity": {
            "minimum": 0.0,
            "title": "Available Quantity",
            "type": "integer"
          },
          "id": {
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Id",
            "type": "string"
          },
          "mapping": {
            "$ref": "#/components/schemas/SupplierMapping"
          },
          "name": {
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Name",
            "type": "string"
          },
          "price": {
            "$ref": "#/components/schemas/Money"
          }
        },
        "required": [
          "id",
          "name",
          "mapping",
          "price",
          "available_quantity"
        ],
        "title": "CatalogVariant",
        "type": "object"
      },
      "FreshnessStatus": {
        "description": "Whether a snapshot is within its declared maximum age.",
        "enum": [
          "fresh",
          "stale"
        ],
        "title": "FreshnessStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MappingApproval": {
        "discriminator": {
          "mapping": {
            "approved": "#/components/schemas/ApprovedMappingApproval",
            "pending": "#/components/schemas/PendingMappingApproval"
          },
          "propertyName": "status"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/PendingMappingApproval"
          },
          {
            "$ref": "#/components/schemas/ApprovedMappingApproval"
          }
        ]
      },
      "Money": {
        "additionalProperties": false,
        "description": "An integer amount in an explicitly named currency and storage unit.",
        "properties": {
          "amount_minor": {
            "minimum": 0.0,
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "unit": {
            "const": "minor",
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "amount_minor",
          "currency",
          "unit"
        ],
        "title": "Money",
        "type": "object"
      },
      "PendingMappingApproval": {
        "additionalProperties": false,
        "description": "A proposed supplier mapping that is not active.",
        "properties": {
          "status": {
            "const": "pending",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "PendingMappingApproval",
        "type": "object"
      },
      "SupplierCapabilities": {
        "additionalProperties": false,
        "description": "Read capability truth with all money and delivery operations locked off.",
        "properties": {
          "catalog_detail": {
            "$ref": "#/components/schemas/Capability"
          },
          "catalog_read": {
            "$ref": "#/components/schemas/Capability"
          },
          "delivery": {
            "$ref": "#/components/schemas/Capability"
          },
          "payment": {
            "$ref": "#/components/schemas/Capability"
          },
          "purchase": {
            "$ref": "#/components/schemas/Capability"
          },
          "refund": {
            "$ref": "#/components/schemas/Capability"
          },
          "top_up": {
            "$ref": "#/components/schemas/Capability"
          }
        },
        "required": [
          "catalog_read",
          "catalog_detail",
          "purchase",
          "payment",
          "top_up",
          "refund",
          "delivery"
        ],
        "title": "SupplierCapabilities",
        "type": "object"
      },
      "SupplierMapping": {
        "additionalProperties": false,
        "description": "A reviewed association between one normalized and one supplier variant.",
        "properties": {
          "approval": {
            "$ref": "#/components/schemas/MappingApproval"
          },
          "supplier_product": {
            "$ref": "#/components/schemas/SupplierProductIdentity"
          },
          "supplier_variant": {
            "$ref": "#/components/schemas/SupplierVariantIdentity"
          }
        },
        "required": [
          "supplier_product",
          "supplier_variant",
          "approval"
        ],
        "title": "SupplierMapping",
        "type": "object"
      },
      "SupplierProductIdentity": {
        "additionalProperties": false,
        "description": "The supplier's opaque identity for a product.",
        "properties": {
          "supplier_id": {
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Supplier Id",
            "type": "string"
          },
          "supplier_product_id": {
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Supplier Product Id",
            "type": "string"
          }
        },
        "required": [
          "supplier_id",
          "supplier_product_id"
        ],
        "title": "SupplierProductIdentity",
        "type": "object"
      },
      "SupplierVariantIdentity": {
        "additionalProperties": false,
        "description": "The supplier's separate opaque identity for a variant.",
        "properties": {
          "supplier_id": {
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Supplier Id",
            "type": "string"
          },
          "supplier_variant_id": {
            "minLength": 1,
            "pattern": ".*\\S.*",
            "title": "Supplier Variant Id",
            "type": "string"
          }
        },
        "required": [
          "supplier_id",
          "supplier_variant_id"
        ],
        "title": "SupplierVariantIdentity",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Nyan Shop Bot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/capabilities": {
      "get": {
        "operationId": "capabilities_api_v1_capabilities_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "all",
                    "mock",
                    "khommo",
                    "vietshare"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierCapabilities"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Capabilities",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog": {
      "get": {
        "operationId": "list_catalog_api_v1_catalog_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "all",
                    "mock",
                    "khommo",
                    "vietshare"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Catalog",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/v1/catalog/{product_id}": {
      "get": {
        "operationId": "catalog_detail_api_v1_catalog__product_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "pattern": ".*\\S.*",
              "title": "Product Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "mock",
                    "khommo",
                    "vietshare"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Catalog Detail",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}